Refactor name of command packages to specific commands names
