+++ bot/messagecommands/summarize/summarize.go
@@ -1,4 +1,4 @@
-package messagecommands
+package summarize
 
 import (
 	"context"
@@ -11,17 +11,6 @@ import (
 	"github.com/sashabaranov/go-openai"
 	"github.com/spf13/viper"
 )
-
-// Send a test message as a response to "jarvis test"
-func Test(s *discordgo.Session, m *discordgo.MessageCreate) {
-	_, err := s.ChannelMessageSend(m.ChannelID, "Hello!")
-	if err != nil {
-		logger.Errorf("Could not send test message response!")
-		return
-	}
-
-	logger.Infof("Successfully sent response to test command to user %s", m.Author.GlobalName)
-}
 
 // The same as Summarize() except for messages before a certain message id.
 func SummarizeBeforeMessageID(s *discordgo.Session, m *discordgo.MessageCreate, beforeMessageID string) {
